Basic recording class

+++ app/src/main/java/net/teamsv/selfalarm/database/Alarm.java
@@ -6,7 +6,7 @@ import io.realm.annotations.PrimaryKey;
 public class Alarm extends RealmObject {
 
     @PrimaryKey
-    private long id;
+    private int id;
 
     private int hour;
     private int minute;
@@ -14,14 +14,14 @@ public class Alarm extends RealmObject {
     private boolean onOff;
 
     /* set funcs */
-    public void setId(long in) { id = in; }
+    public void setId(int in) { id = in; }
     public void setTime(int h, int m) { hour = h; minute = m; }
     public void setRecord(Record in) { record = in; }
     //public void setRecord() { } // Select random alarm
     public void setOnoff(boolean in) { onOff = in; }
 
     /* get funcs */
-    public long getId() { return id; }
+    public int getId() { return id; }
     public int getHour() { return hour; }
     public int getMinute() { return minute; }
     public String getTime() {
